Link ethashcl also if it's found in the path
In windows when building with split repositories webthree seems to
require ethashcl if first libethereum was built with libethash-cl
on (which is the default)

The way webthree's cmake files are at the moment it will try to link to
ethashcl but because nowhere is `ETHASCL` set it will not do it,
resulting in undefined link symbols.

With this commit I believe we also allow linking not only if ETHASHCL has
been set but also if  we have already compiled ETHASH-CL and cmake found it.